fix: sync old data -  fixed the comparison between the dates of  the globalState data and today's date - updated the createDailyData and updateDailyData of the daily-data service to use the dates passed in parameter and no longer today's date (which leaded to fk errors) -  updated the dtos in the daily-data.dto to properly pass the targetedDate down
